original code

+++ libtommath/bn_error.c
@@ -1,4 +1,4 @@
-#include "tommath_private.h"
+#include <tommath_private.h>
 #ifdef BN_ERROR_C
 /* LibTomMath, multiple-precision integer library -- Tom St Denis
  *
@@ -9,28 +9,31 @@
  * Michael Fromberger but has been written from scratch with
  * additional optimizations in place.
  *
- * SPDX-License-Identifier: Unlicense
+ * The library is free for all purposes without any express
+ * guarantee it works.
+ *
+ * Tom St Denis, tstdenis82@gmail.com, http://libtom.org
  */
 
 static const struct {
-   int code;
-   const char *msg;
+     int code;
+     const char *msg;
 } msgs[] = {
-   { MP_OKAY, "Successful" },
-   { MP_MEM,  "Out of heap" },
-   { MP_VAL,  "Value out of range" }
+     { MP_OKAY, "Successful" },
+     { MP_MEM,  "Out of heap" },
+     { MP_VAL,  "Value out of range" }
 };
 
 /* return a char * string for a given code */
 const char *mp_error_to_string(int code)
 {
-   size_t x;
+   int x;
 
    /* scan the lookup table for the given message */
-   for (x = 0; x < (sizeof(msgs) / sizeof(msgs[0])); x++) {
-      if (msgs[x].code == code) {
-         return msgs[x].msg;
-      }
+   for (x = 0; x < (int)(sizeof(msgs) / sizeof(msgs[0])); x++) {
+       if (msgs[x].code == code) {
+          return msgs[x].msg;
+       }
    }
 
    /* generic reply for invalid code */
@@ -39,6 +42,6 @@ const char *mp_error_to_string(int code)
 
 #endif
 
-/* ref:         HEAD -> master, tag: v1.1.0 */
-/* git commit:  08549ad6bc8b0cede0b357a9c341c5c6473a9c55 */
-/* commit time: 2019-01-28 20:32:32 +0100 */
+/* ref:         $Format:%D$ */
+/* git commit:  $Format:%H$ */
+/* commit time: $Format:%ai$ */
